Use universal she-bang for Debian-based systems
On some versions of Debian (and thus, Ubuntu and friends), `node` is installed as `nodejs`.  This means that the default node shebang does not work, since there is no `node` binary.  This shebang runs as shell, and then invokes the remainder of the script as node or nodejs, depending on what's installed on the system.

Because this happens in the shebang, it is outside of the `try / catch`, and therefore fails any git operations, regardless of if there are hooks installed or not.